ExecutionParametersDialog: Changes cast from int to float in shifttfs
Former-commit-id: 0bcb0b926b63d6cdb87a4b2ec6dd9109c92b327a

+++ mod/dataobjects/execution_parameters.py
@@ -21,8 +21,8 @@ class ExecutionParameters():
         self.densitydt_type: DensityDTType = DensityDTType.NONE
         self.densitydt_value: float = 0.1
         self.shifting: int = 0
-        self.shiftcoef: int = -2
-        self.shifttfs: int = 0
+        self.shiftcoef: float = -2
+        self.shifttfs: float = 0
         self.rigidalgorithm: int = 1
         self.ftpause: float = 0.0
         self.coefdtmin: float = 0.05
